osa4-kaikki-tehtavat

+++ osa4/osa4-blogilista/controllers/blogs.js
@@ -1,29 +1,51 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+
 
 blogsRouter.get('/', async (request, response) => {
-  const posts = await Blog.find({})
+  const posts = await Blog.find({}).populate('user', {id: 1, username: 1, name: 1} )
   response.json(posts)
 })
 
 blogsRouter.post('/', async (request, response) => {
+  const body = request.body
   try {
-    const body = request.body
+    
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!request.token || !decodedToken.id) {
+      return response.status(401).json({error: 'token missing or invalid'})
+    }
+
     if(body.title === undefined && body.url === undefined ) {
       return response.status(400).json({ error: 'content missing' })
     }
+
+    const user = await User.findById(decodedToken.id)
+
     const blogpost = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes === undefined ? 0 : body.likes
+      likes: body.likes === undefined ? 0 : body.likes,
+      user: user._id
     })
     const savedBlogpost = await blogpost.save()
+
+    user.posts = user.posts.concat(savedBlogpost._id)
+    await user.save()
     //console.log(savedBlogpost)
-    response.json(blogpost)
+    response.json(Blog.format(blogpost))
   } catch (exception) {
-    //console.log(exception)
-    response.status(500).json({ error: 'something went wrong...' })
+    if (exception.name === 'JsonWebTokenError') {
+      response.status(401).json({error: exception.message})
+    } else {
+      //console.log(exception)
+      response.status(500).json({ error: 'something went wrong...' })
+    }
   }
 })
 
@@ -46,11 +68,30 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
-    const post = await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    
+    if(!request.token || !decodedToken.id) {
+      return response.status(401).json({error: 'token missing or invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    const user = await User.findById(decodedToken.id)
+
+    if (blog.user.toString() === user.id.toString()) {
+      const post = await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).json({error: 'token invalid'})
+    }
+  
+    
   } catch (exception) {
-    console.log(exception)
-    response.status(400).send({error: 'cannot find id'})
+    if (exception.name === 'JsonWebTokenError') {
+      response.status(401).json({error: exception.message})
+    } else {
+      console.log(exception)
+      response.status(400).send({error: 'cannot find id'})
+    }
   }
 })
 
@@ -65,7 +106,7 @@ blogsRouter.put('/:id', async (request, response) => {
       likes: body.likes
     }
     const updatedpost = await Blog.findByIdAndUpdate(request.params.id,updatedBlogPost, { new: true})
-    response.json(updatedpost)
+    response.json(Blog.format(updatedpost))
   } catch (exception) {
     //console.log(exception)
     response.status(400).send({error: 'malformatted id'})
